MSR: add support for verifying a written register value
Add a new parameter 'verify' to the MSR write APIs, which, if set
to True, will read-back the written value and verify that it gets
properly through. If the verification fails, an ErrorVerifyFailed
exception is raised. By default, none of the writes are verified.
Also, please note that using the 'verify' will flush the existing
transaction cache if it is in use.

Signed-off-by: Tero Kristo <tero.kristo@linux.intel.com>