_FeaturedMSR: fix yield bug
There is some weird bug in Python; when a function can yield and return nothing,
even when the yield is never reached.

In this case 'write_feature()' had a yield that was never used, but because of
its existence some MSR values were not written and no exception raised. Only
when a test read back the value and compared it to the written value was the
bug noticeable.

Feels like a memory leak in C, the side effects usually appear in random places.

Signed-off-by: Niklas Neronin <niklas.neronin@intel.com>