MSR: fix verification logic to work properly with emulation
In current 'verify' logic implementation the emulated read-back only sees
the proper value for the first CPU register modified, rest get invalid
values. This is caused by the improper handling of cached data, where we
remove the cached value first for each individual CPU, and then
read-back the value. To fix the problem, remove the cached data for all CPUs
before attempting to read-back any value, and after that rely on the cached data.

Signed-off-by: Tero Kristo <tero.kristo@linux.intel.com>