ClassHelpers: cleanup naming in 'WrapExceptions'
Some attributes of 'WrapExceptions' class have the '_we_' prefix, some do not.
This is inconsistent. The original idea behind the '_we_' prefix was to avoid
possible name collisions with attributes of the wrapped class. But this is not
a real problem anyway. So just clean-up the class and remove the '_we_'
prefixes.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>