test_pstates_cmd: fix invalid 'pinfo' format in 'test_pstates_save_restore()'
This is a bug that affects systems/datasets where not all CPUs have the
same property value. The bug is hard to describe so I have written a
step by step guide to how/why it failed.

This is how a 'pinfo' might look for property 'min_freq':
  [{"value" : 1000000000, "cpus" : "0"},
   {"value" : 1500000000, "cpus" : "1-128"}]
As you can see different values have their own dictionary in the list.

Before, in 'test_pstates_save_restore()' we would set all 'min_freq'
values to the same value, like this:
  [{"value" : 1000000000, "cpus" : "0"},
   {"value" : 1000000000, "cpus" : "1-128"}]
This was then used to restore the values.

Afterwards we would read and save the values to compare them, but now 'pinfo'
looks different:
  [{"value" : 1000000000, "cpus" : "0-128"}]

This would make the "read_back == state_modified" assert fail. Thus, when
setting the values we create a new list with one dictionary:
  [{"value" : 1000000000, "cpus" : "0-128"}]

Signed-off-by: Niklas Neronin <niklas.neronin@intel.com>