PStates: extend freq. set smart messages to min_freq
When we set frequency to a value via the sysfs file, then read the file back
and the value is different to what we wrote, we try to provide a smart message,
and there is a quirk for AMD systems.

The smartness and the quirk, however, work only for the 'max_freq' property and
do not cover the 'min_freq' property. As a result, a command like

pepc pstates config --min-freq max

will not get the smart message. Fix this by also covering the 'min_freq'
property.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>