tdgen: fix handling of 0 byte files
Inline file line format: "<path to file>:<file contents>"

The previous approach relied that all files would always contain a new
line, this was important because no new line wa added by tdgen.
This caused issues when writing the contents to an inline file, because
of no new line 2 or more files were writen on the same line, example:
  /path/file:/other/path:123
instead of:
  /path/file:
  /other/path:123

In this patch when writing inline files a new line will be added after ":".
 '.*' - forces to print each file, even if it is 0 bytes.
 '-H' - prepends file path, "/tmp/Hello.txt:Hello".

Signed-off-by: Niklas Neronin <niklas.neronin@intel.com>