tdgen: fix copying '/proc/cpuinfo'
'/proc/cpuinfo' can not be copied directly using 'scp' or 'rsync', in some
cases only a part of the file is copied. To circumvent this issue, we read
the file with 'cat' on the remote host, and then send the data to the local
host, which writes it to a file.

Signed-off-by: Niklas Neronin <niklas.neronin@intel.com>