_PepcCommon: add function to override 'CPUInfo' CPU model
This function will be used in the next commit to override CPU model
when option '--override-cpu-model' is used.

In practice this means that the user can run on CPU A is if it is CPU B,
even if CPU A is not supported by Pepc. The argument can be an integer,
or string representation of an decimal/hex.

Signed-off-by: Niklas Neronin <niklas.neronin@intel.com>