pepctools: add option '--override-cpu-model' to 'power'
This options can be used to override the CPU model. For example, the users
current CPU is not supported by Pepc, but it is similar to another CPU which is
supported by Pepc, using '--override-cpu-model' the user can override the CPU
model to the supported CPU and then run 'power info/config'.

Signed-off-by: Niklas Neronin <niklas.neronin@intel.com>