_Pepc: remove all emulated data if some data is missing
===== Issue =====
In some cases emulated data was left in '/tmp/' after the test had run.

===== Cause =====
For some tests we don't yet have emulated data ('power'), and after failing to
create the emulated data, the existing emulated data in '/tmp' (e.g., CPUInfo)
was not removed.

===== Breakdown =====
In the example below we have 'get_pman()' that returns a 'pman' object. In the
case of 'get_pman()' raising 'ErrorNotSupported' we would catch it and skip the
test. But, because of the exception, 'with' does not close 'pman', and any
emulated data that was created before the exception is not removed.

try:
    with get_pman(hostspec, modules=emul_modules) as pman:
        ...
except ErrorNotSupported:
    pytest.skip("Data not yet available for platform.")

===== Fix =====
In this patch we catch 'Error' raised by  'init_testdata()' and close
'EmulProcessManager' object.

===== Notes =====
This fixes the issue that we currently have, but if 'get_pman()' would be
structured like the example below, the emulated data would not get removed.
But I believe this a 'init_testdata()' user problem.

def get_pman():
    ...
    pman = ProcessManager.get_pman(hostspec, username=username)

    if datapath and modules is not None:
        pman.init_testdata(modules, datapath)

    raise Error("Some function raises error etc.")
    return pman

Signed-off-by: Niklas Neronin <niklas.neronin@intel.com>