pepc: improve help text even more
Add new key to properties - "mechanisms", which provides the mechanisms the
property is accessed. Currently only 'sysfs' and 'msr' mechanisms are
supported.

With this new key, we can remove the "via MSR" and "via sysfs" bits from
property names, which makes it possible to format better help text.

Compare before and after.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>