msrscope: add 'PCStateConfigCtl' MSR scope tests
'pkg_cstate_limit' is an edge case, because when writing the property takes a
package cstate (e.g., "PC2", "PC0"). But when reading the feature does not
return a package cstate, rather a dictionary with a "pkg_cstate_limit" entry
containing the package cstate limit. Thus, '_find_feature_scope' had to be
adjusted.

Signed-off-by: Niklas Neronin <niklas.neronin@intel.com>