PowerCtl: set die scope for Cascade Lake AP
MSR_POWER_CTL features have package scope, except for Cascade Lake AP which has
die scope, because Cascade Lake AP has multiple dies per package. Thus, we set
the scope to die on CPU models 85 with multiple dies per package. This is not a
perfect fix, if all die 1 CPUs are offline before the module is initialized then
the scope will be set to package. In practice this is a very rear occurrence,
and its eventuality everything will work as if package scope is the correct
scope.

Signed-off-by: Niklas Neronin <niklas.neronin@intel.com>