msrscope: optimize by writing and reading only one value
'msrscope' requires 2 unique values, the MSR is then set to each of the
values. This prevents the value being set to the same value.

To optimize performance, instead of setting both values, we only use
the value which is different from the value already set in the MSR.

For example, 'msrscope 0x1A4 --values 1 0':
'1' will be skipped because the MSR is already '1'.
'0' will be used, because the MSR value is '1'.

Signed-off-by: Niklas Neronin <niklas.neronin@intel.com>