_PepcTopology: add '--hybrid' option
This option will include E/P-core information, see example below:

./pepc topology info --hybrid
CPU    Core    Module    Node    Package    Hybrid
  0       0         0       0          0    E-core
  1       1         0       0          0    E-core
  2       2         0       0          0    P-core
  3       3         0       0          0    P-core

In case the CPU is not hybrid, the tool will warn the user and omit
"Hybrid" column.

The reason why "Hybrid" is its own option, is because it does not follow
the same rules as core, node, package etc. For instance, the sysfs files
that contain the hybrid CPU topology only exist on hybrid CPUs.
Additionally, the value ("E-core"/"P-core") is a string and not an integer,
so we can't sort them in ascending order.
Due to these differences "Hybrid" column is its own option, this makes
it clear that '--order' and '--column' will not work with the "hybrid"
argument.

Signed-off-by: Niklas Neronin <niklas.neronin@intel.com>