Pepc: remove sub-property functionality
What is the difference between properties and sub-properties?
1. Sub-properties are read-only.
2. When a property is not supported, all its sub-properties are assumed to
   be not supported.

For #1 we already have an alternative; "writable" key. Which when set to
'False' makes the property read-only.

For #2 we don't have anything equivalent. But there are currently only 2
properties that have sub-properties; 'pkg_cstate_limit' and 'governor'.
Property 'pkg_cstate_limit's sub-properties, except for
'pkg_cstate_limit_lock', are actually aliases, so their supported status
mirrors their "main" property. And as for 'governor' its sub-property
'governors' can not be read, if 'governor' can not be read.

As you can see sub-property functionality is not very unique and does
not justify its complexity. The only end user visible difference from
this patch is that on system where 'pkg_cstate_limit' is not supported,
'pkg_cstate_limit_lock' may still say "on/off".

Additionally, this is 'pinfo's current format:

  { property1_name : { property1_name : value },
    property2_name : { property2_name : value,
                       property2_subprop1_name : value,
		       ... and so on of all sub-properties ...
		     },
    ... and so on for all properties ...
  }

When sub-property functionality is removed, 'pinfo's format can then be
simplified:

  { property1_name : value,
    property2_name : value,
    property3_name : value,
    ... and so on for all properties ...
  }

Signed-off-by: Niklas Neronin <niklas.neronin@intel.com>