_PepcPrinter: fix 'print_props()' argument 'skip_ro'
'pepc_props()' docstring states that argument 'skip_ro' concerns properties
and sub-properties, but in reality it only affects sub-properties.

So the question is: Is the docstring incorrect or is it a bug?

Firstly I checked the commit that added 'skip_ro'
 - It was added in commit: f0696c4 "pepc save/restore: redesign", the same
   issue exist in the original commit.
Secondly I checked who uses 'skip_ro=True' argument.
 - The argument is used by save and restore functionality in 'PStates',
   'CStates' and 'Power'. Save and restore is only for properties that
   are writable. Thus, the docstring is correct and it's a bug.

Signed-off-by: Niklas Neronin <niklas.neronin@intel.com>