_PepcPrinter: remove unnecessary exception
Exception that 'action' must be 'None' when printing in YAML format, is
unnecessary. When format is "yaml", 'action' is ignored, nothing will
break/crash if 'action' is not 'None' when printing to YAML.

Function 'print_props()' does not have "format" argument, and is a
generic "print properties" function. Thus, it should not raise format
specific exceptions, especially they do not break anything.
The exception would be appropriate if the function was instead called:
'print_props_yaml()'.

Signed-off-by: Niklas Neronin <niklas.neronin@intel.com>