_PepcPrinter: optimize 'print_props()'
Don't convert "all" to list of online CPUs.

Function 'PropsClassBase.get_props()' will normalize the 'cpus' argument.
When doing so, if the value is "all" it returns list of online CPUs which
is cached in 'CPUInfo'. But if 'cpus' is a list of CPUs, it will check if
each CPU exists in the cached online CPUs list, which is much slower.

Signed-off-by: Niklas Neronin <niklas.neronin@intel.com>