_PepcPrinter: rename and massage human printing functions
Rename '_print_aggr_pinfo_group()' to '_do_print_aggr_human()', because it will
be used for pringint non-grouped properties, so "group" part in the name is
removed. And it is a helper for '_print_aggr_pinfo_human()', therefore call it
the same way, juat add a '_do_' prefix.

Move hard-coded " - " prefix to the main function too. This way
'_do_print_aggr_pinfo_human()' becomes more flexible and independent on
grouping.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>