_PepcPrinter: clean-up aggregate properties info dictionary
Backgraound
-----------

When the aggregate properties dictionary ('aggr_pinfo') is built for
'pepc cstates save', all sub-properties are excluded. Reason: they they are
read-only and should not be saved, because they can't be restored.

But there is one exception - the 'pkg_cstate_limit_locked' sub-property. It is
included to the aggregate dictionary. Reason: if the package C-state limit MSR
is locked, it is impossible to restore package C-state limit, so it should not
be saved.

In other words, the aggregate dictionary excludes all sub-properties, except for
'pkg_cstate_limit_locked'.

Problem
-------

The main task of the '_adjust_aggr_pinfo_pcs_limit()' function is to remove
locked CPUs from the aggregate dictionary. If only some CPUs are locked and
removed, this function also deletes the 'pkg_cstate_limit_locked' key. Hoerver,
if the package C-state limit property is not even supported, this function does
not delete the 'pkg_cstate_limit_locked' key.

This behavior is inconsistent. It does not lead to bugs at this point, but it
may in the future if something around 'aggr_pinfo' handling gets modified. It
is cleaner and less error-prone to always delete the "locked" key.

Solution
--------

1. Move 'pkg_cstate_limit_locked' key deletion to 'print_props()' - the same function
   where it was "added" at. This is just simpler and easier to understand:
   a. decide to keep 'pkg_cstate_limit_locked' in 'print_props()
   b. call '_adjust_aggr_pinfo_limit()' to remove the locked CPUs
   c. remove 'pkg_cstate_limit_locked' in 'print_props()'
2. While on it, also move 'pkg_cstate_limit_locked' key deletion to
   'print_props()' for consisdtency.
2. Make sure 'pkg_cstate_limit_locked' is always removed.

Additional change
-----------------

In addition to the 'pkg_cstate_limit_locked' change, move deletion of the
'pkg_cstate_limit' key to 'print_props()' method. This way the split of labor
becomes clean and consistent.

1. print_props():
   a. forms aggregate properties dictionary.
   b. makes decision to keep the "locked" sub-property
   c. uses '_adjust_aggr_pinfo_pcs_limit()' to remove locked CPUs
   d. cleans up the aggregate dictionary by removing all unneeded keys.
2. _adjust_aggr_pinfo_pcs_limit():
   a. Focuses only on removing locked CPUs
   b. Does not remove any keys, which makes it simpler

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>