_PepcPrinter: fix '_adjust_aggr_pinfo_pcs_limit()'
This statement does not work correctly:
 - if not pcsl_info or not lock_info

'pcsl_info' always contains a dictionary with at least 1 key.
'lock_info' always contains a dictionary with at least 1 key if 'pcsl_info'
contains a key that is not 'None'.

If a property is not supported, then the "info" is:
 - '{None : [ list of at least one CPU ]}'
Thus, instead of checking "not pcsl_info", check if 'pcsl_info.keys'
contains only one key which is 'None':
 - if pcsl_info.keys() == {None, }

There is not reason to check if 'pkg_cstate_limit_locked' keys are
'{None, }', because we proceed only if 'pkg_cstate_limit_locked'
contains a "on" key.

Signed-off-by: Niklas Neronin <niklas.neronin@intel.com>