_PepcPrinter: optimize '_adjust_aggr_pinfo_pcs_limit()' when no CPUs are locked
The aggregate properties information dictionary 'aggr_pinfo' includes
'pkg_cstate_limit' property. This property is read/write in case the
corresponding MSR is unlocked, and it is R/O if the MSR is locked.
The goal of this method is to remove all the "locked" CPUs from the
'pkg_cstate_limit' key of 'aggr_pinfo'.

So when all CPUs are unlocked, there is nothing to do so return 'aggr_pinfo'
without modification.

Signed-off-by: Niklas Neronin <niklas.neronin@intel.com>