_PepcPrinter: optimize '_adjust_aggr_pinfo_pcs_limit()' when all CPUs are locked
The aggregate properties information dictionary 'aggr_pinfo' includes
'pkg_cstate_limit' property. This property is read/write in case the
corresponding MSR is unlocked, and it is R/O if the MSR is locked.
The goal of this method is to remove all the "locked" CPUs from the
'pkg_cstate_limit' key of 'aggr_pinfo'.

When length of locked CPUs is the same as all requested CPUs, this can only mean
that all requested CPUs are locked. So there is no reason to check each CPU,
'pkg_cstate_limit' is considered as read-only and is be removed.

Signed-off-by: Niklas Neronin <niklas.neronin@intel.com>