_PepcPrinter: don't print C-state if it is missing on/off info
No point in printing information about a C-state when enabled/disabled status
is unknown. Otherwise, as you can see from the example below, sub-values
"description", "expected latency" and "target residency" are printed twice,
the second time is for another C-state, not "POLL".

Before:

POLL: 'on' for CPUs 0-191 (all CPUs)
 - description: CPUIDLE CORE POLL IDLE
 - expected latency: 0 us
 - target residency: 0 us
 - description: MWAIT 0x00
 - expected latency: 2 us
 - target residency: 2 us

After:

POLL: 'on' for CPUs 0-191 (all CPUs)
 - description: CPUIDLE CORE POLL IDLE
 - expected latency: 0 us
 - target residency: 0 us

It could be changed to only print the C-state name, but the sub-values are not
useful without knowing if C-state is enabled/disabled.

Signed-off-by: Niklas Neronin <niklas.neronin@intel.com>