pepc cstates config: never print source
Fix the following issue:

$ pepc cstates config --disable all --enable POLL -H bdwep1
Source: Linux sysfs file-system
 - POLL: set to 'off' for CPUs 0-87 (all CPUs)
 - C1: set to 'off' for CPUs 0-87 (all CPUs)
 - C1E: set to 'off' for CPUs 0-87 (all CPUs)
 - C3: set to 'off' for CPUs 0-87 (all CPUs)
 - C6: set to 'off' for CPUs 0-87 (all CPUs)
POLL: set to 'on' for CPUs 0-87 (all CPUs)

The "Source:" part should not be there. Fix the output and make it look like
this:

```
$ pepc cstates config --disable all --enable POLL
POLL: set to 'off' for CPUs 0-87 (all CPUs)
C1: set to 'off' for CPUs 0-87 (all CPUs)
C1E: set to 'off' for CPUs 0-87 (all CPUs)
C3: set to 'off' for CPUs 0-87 (all CPUs)
C6: set to 'off' for CPUs 0-87 (all CPUs)
POLL: set to 'on' for CPUs 0-87 (all CPUs)

The problem is that grouping decision is made in the '_PropsPrinter' class
based on whether all propeties are printed or not all of them. This is not the
ideal way to make the decision. Instead, the decision should be made outside of
'_PropsPrinter', by the user of the class. The class should just follow what
the user said.

Move the 'group' argumet from internal '_PropsPrinter' methods to its public
'print_props()' and 'print_cstates()' methods. This way, the caller can make
the grouping decision.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>