_PropsClassBase: add 'mnames' argument to 'set_props()'
This is a preparation for imroving properties API to support writing propertis
using a specific, user-defined mechanism.

Add the 'mnames' argument to 'set_prop()' and 'set_cpu_prop()' methods. This
option will be used for specifying what mechanism should be used for setting
the property. Do not implement it yet.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>