_PropsCache: rework 'get()' and 'is_cached()'
This patch aims to optimize 'get()' and 'is_cached()', and prevent 'get()'
failing when caching is disabled.

Before, 'is_cached()' had to be called before 'get()', because
'is_cached()' validated that caching is enabled and if the value is
cached, afterwards 'get()' could retrieved the value. This meant that if
caching was disabled and 'get()' was called, an 'AttributeError' was
raised. Additionally, the value was retrieved twice, once in
'is_cached()' and once in 'get()'.

Signed-off-by: Niklas Neronin <niklas.neronin@intel.com>