PStates: add 'set_freq_props()' method
Add a new method - 'set_freq_props()'. It is a special case for 'set_prop()',
but it accepts min. and max. frequencies. This method handles the min. and max.
frequency set oreder.

This method is very similar to '_validate_and_set_freq()', which will go away
soon, along with 'set_props()' method.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>