CPUInfo: turn warning to a debug messages
When 'pepc' runs on a non-hybrid system with old kernel, it warns about hybrid
system support only in v5.13, which does not make sense. We'd also need to
check if the system is generally hybrid (e.g., via CPUID), but we do not do
this.

Therefore, turn the warning into a debug message.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>