_PropsClassBase: change 'get_prop()' API
Change 'get_prop()' and 'get_cpu_prop()' return values. Instead of a tuple,
return a dictionary, which is referred to as "pvinfo". Put CPU number and
property value to that dictionary.

This is preparation for upcoming changes where 'get_prop()' will also return
the mechanism name that was used for reading the property.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>