_PerCPUCache: new module
Currently '_PropsCache' is designed for property modules, it even includes the
'mname' argument. But '_PropsCache' is also used from 'CPUIdle' and 'MSR'
modules, which have nothing to do with properties. This is conceptually wrong.

Create a simplified version of '_PropsCache' which just caches any per-CPU
data, and does not know about properties and mechanisms. Use it from 'CPUIdle'
and 'MSR'.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>