_PropsClassBase: add per-package and per-die interfaces
Chage the properties interface to work not only on per-CPU, but also on
per-package and per-die basis. That is, allow the caller to get and set
properties for the entire die or the entire package by specifying die and
package numbers. Contrast this with current implementation where users have to
specify specifying CPU numbers, because the API only allows for CPU numbers.

Add 'get_prop_dies()' and 'get_prop_packages()' methods for getting a property
on per-die and per-package basis.

Add 'set_prop_dies()' and 'set_prop_packages()' methods for setting a property
on per-die and per-package basis.

Do not implement teh new methots for real yet, just emulate for now.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>