PCStateConfigCtl: rename "lock" to "pkg_cstate_limit_lock"
The "lock" feature is the only MSR feature not matching the corresponsing
property name. This is inconsistent and leads to a subtle bug in '_set_sname()'
which does not set scope name for "pkg_cstate_limit_lock", because
'if pname in PCStateConfigCtl.FEATURES:' is 'False' for "lock".

Fix this make code more consistent by renaming the "lock" feature to
"pkg_cstate_limit_lock".

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>