PStates: _CPUFreq: separate out CPU frequency operations, step 1
This is the first change in a series of changes which have the end goal of
separating out CPU frequency operations from 'PStates' to a new a module, in
order to make 'PStates.py' be smaller and more managable.

Create new module - '_CPUFreq', which is similar to '_UncoreFreq'. Add the
'CPUFreqSysfs' class and implement the following methods:
  * 'set_min_freq()'
  * 'set_max_freq()'

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>