Logging: return resolved log path from 'setup_stdout_logging()'
This patch adjusts 'setup_stdout_logging()' so that it returns the
resultant path to which logs will be written.

This change will allow functions/classes which use this function to
reference the log file.  For example, this would allow a function to
present the path of the log file to the user so that they can manually
inspect it.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>