_OpTarget: new class representing an operation target
Add a new 'OpTarget' class representing the target of a pepc operation, meaning
the CPUs, cores, modules, dies, or packages the operation should be applied to.

Partially switch the "pepctool"-level modules to using the new class. This
makes class encapsulates the CPU/package/etc numbers and operations on them,
which will make the implementation cleaner and easier to maintain.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>