PepcPylintPlugin: relax function argument parsing to accept "same as"
If a string "same as '<func-name>()'" is detected in function docstring,
check that the documentation for the undocumented arguments is covered
by the referred function. Also, add support for more stricter check that
can be optionally enabled, which complains if the referred function
can't be found from the source file currently being parsed.

Signed-off-by: Tero Kristo <tero.kristo@linux.intel.com>