PStates/CStates: improve '_set_sname()'
When setting property scope name fails for some reason, do not assign the "CPU"
scope, let the method raise an exception. Assigning the "default" scope may
lead to hiding bugs. Instead, expose the fact that the property is not
supported to the caller, and let the caller deal with this.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>