_PepcPrinter: change aggregate properties information dictionary format
Prepare the printer for the upcoming changes where the property will be read on
per-CPU, per-die, or per-package basis, depending on its scope. For these
purposes, the aggregate information dictionary should be changes so that it
carries the scope name and the numbers are not necessarily CPU numbers, but may
also be die or package numbers.

Re-structure the aggregate information dictionary correspondingly. Do not
implement per-die or per-package reading yet, just prepare the data structure
to be ready.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>