_PropsClassBase: add I/O scope name to properties information dictionary
Add the I/O scope name to the 'self._props' properties information dictionary.
It is the same as scope name for everything except for the package C-state
limit today.

This change is a preparation to the upcaming changes where properties with I/O
scope smaller than scope will be additionally validated.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>