_PropsClassBase: catch incorrectly configured systems
When property scope is not the same as its I/O scope, it is possible to
mis-configure the system so that it becomes hard to determine what is the real
value of the property.

Add logic for catching such situations.

Example one a system with one die per package:

$ pepc cstates info --pkg-cstate-limit --package 3
pepc: error: cannot determine the value of property 'pkg_cstate_limit' for package 3:
  CPU 120 has value 'PC6R', but CPU 159 has value 'PC0', even though they are in the same package.
  This situation is possible because 'pkg_cstate_limit' has 'package' scope, but 'core' I/O scope.

Example for a multi-die system:

$ pepc cstates info --pkg-cstate-limit --package 1
pepc: error: failed determine the value of property 'pkg_cstate_limit' for package 1, die 1:
  CPU 72 has value 'PC6R', but CPU 95 has value 'PC0', even though they are in the same die.
  This situation is possible because 'pkg_cstate_limit' has 'die' scope, but 'core' I/O scope.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>