CPUInfo: fix topology for I/O dies
Based on work by Tero Kristo <tero.kristo@linux.intel.com>.

When the topology is being built, the first step is creating a temporary version
of it which is referred to as 'tinfo' and which is a dictionary indexed by CPU
number. However, for I/O dies there are no CPUs, so all I/O dies end up being
reduced to a single dictionary entry with key '_NA'. As a result, the final
topology loses all I/O dies except for one.

Fix this by always appending I/O dies to the topology table at the end. Either
get them by scanning the target system (when done for the 1st time) or from the
I/O dies cache.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>