_SysfsIO: new module
There are multiple places in the project whre sysfs files are read and written
to, and the result is cached. This is the first step in unifying those places.

Add '_SysfsIO' class, implement 'read()', 'read_int()' and 'write()' methods,
do not implement cachine yet.

Switch the '_CPUFreq' module to using '_SysfsIO'.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>