_PropsClassBase: push 'cpus' handling 1 step down
Thesis: it is beneficial to propagate all CPUs down to low-level methods,
because then it is possible to optimize reading the properties in bulk. When
each operation is done on an individual CPU basis, bulk reads are impossible.

Move one step towards the goal - push the 'cpus' argument handling from
'_get_prop_cpus()' down to '_get_prop_pvinfo_cpus()'.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>