_PropsClassBase: raise 'ErrorNotSupportred' for usupported properties
When a property is not supported, many methods return 'None'. This is very
inconvenient, because it makes it hard to distinguish between the following
situation:

1. The property is not supported at all.
2. The property is not supported for a CPU, but may be supported for another
   CPU.

Change property "get" methods to raise 'ErrorNotSupported' when a property is
not supported at all. Howerver, hide this exception from users of the property
classes so far. Just raise the exceptions internally for now, but translate
them to 'None' if 'pvinfo' dictionaries at the highest level (the base class).

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>