PStates: partially switch '_set_freq_prop()' to multi-CPU methods
Partially switch '_set_freq_prop()' to using multi-CPU methods. Longer term
this will help optimizing the tool by implementing bulk MSR or syfs access in
low level modules like 'MSR' and '_CPUFreq'. For that, they have to be accessed
in a multi-CPU manner. This is a step towards the long term goal.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>