_SysfsIO: introduce 'write_verify_int()'
Add a new helper - 'write_verify_int()', that is similar to 'write_verify()',
but when it raises 'ErrorNotSupported()', it stores the actual and expected
values in the exception objects as integers, as opposed to storing them as
strings.

This, in turn, fixes a crash in 'self._handle_write_and_read_freq_mismatch()',
which expects the frequency to be an integer and crashes otherwise.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>