_PepcPrinter: fix human formatting for lists of numerical values
Fix an issue in '_format_value_human()' which does not properly format
list values containing one float or integer.

If a property value is a list of float/int but the list only contains
one number, it should be formatted as a single value. Fix the
'_format_value_human()' method so that it does this.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>