tests_cpuonline: re-write the test
Re-write the test so that it does not try to preserve online/offline CPU
status, but just onlines everything, tests, and onlines everything at the end,
even if there is an error. This makes it possible to re-run tests after
failures and interruptions.

Also do not try to offline all CPUs - this does not work on most real systems.

And make the test also run on real systems, not only the emoulated systems.

And address all pylint warnings.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>