Tpmi: move spec dirs discovery to the tool level
Spec directory paths discovery does not belong to the library level, it belongs
to the tools level, and should be just the input parameter for the library,
because in theory, different tools may have different directories and different
ways of discovering them.

Move the spec file directories discovery code to '_PepcTpmi.py'.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>