Tpmi: introduce fmap
Introduce a new term - feature map (fmap). This is a dictionary that allows
finding debugfs paths for every known feature.

Right now it just contains the feature debugfs paths list, but later it can be
improved to include package information and more.

The idea is that if one needs to read or write a TPMI register for feature "X",
then 'fmap["X"]' would provide all the necessary information about where to
read or write to.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>