EmulProcessManager: fix 'lsdir()' to return proper pathnames
Currently, 'lsdir()' returns the actual pathnames of the emulated files,
including the temporary directory where they reside. This causes
problems when user then later tries to use the emulated pathname for
some purpose, leading it being prefixed twice with the temporary
directory.

To fix the problem, remove the temporary path prefix from the path that
gets returned to the user.

Artem: massaged to avoid casting to 'str()'.

Signed-off-by: Tero Kristo <tero.kristo@linux.intel.com>
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>