TPMI: improve mdmap and instance definitions
Improve definitions of the "mdmap" and "instance" term. Make it a bit more
detailed. Also, do not call "mdmap" as "memory dump map", because it is
conusing. It is more like "mem_dump file map". But it is better to stick with
only one word for it - mdmap, and avoid using anything else.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>