Tpmi: rename '_get_memdump_map()'
Let's stick with one single term - mdmap - everywhere, so do not use 'memdump'.

Also, prefer using '_get_something()' for methods "that get from a cache or
build if not in cache" sort of methods. Since this method does not cache
anything, just builds and returns, name it '_build_mdmap()' instead.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>