Tpmi: improve '_build_mdmap()' docstring
Shorten the docstring, as we have extensive explanation in the definitions
section. Stick to "mdmap", do not use "memory dump" words at all.

Also, stick with these terms for clarity:
  - offset - always mean the TPMI register offset, comes from spec file.
  - position - always mean the 'mem_dump' file position. Like in Linux,
               we call it 'f_pos', also check 'seek()'.

Aviod using "byte offset". Offset is in bytes, not in words by default. Do not
use offset for 'mem_dump' position. This improves clarity and readability.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>