Tpmi: change the fdict meaning
Re-purpose fdict to only contain the "registers" part of the spec file. The
"header" of the spec file is in sdict, no need to keep it in fdict too.

Add a stub for spec file validation method, which we need to implement.

Introduct '_get_sdict()' helper. Refactor '_get_regdict()' to use the new
semantics of the fdict.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>