Tpmi: refactor '_get_regdict()'
Do not include all register names to the error message - it will be massive and
unreadable. Shorten the error message. The API level methods will have to do
input argument validation and if a register does not exist, they will have to
raise an error. So '_get_regdict()' method should always get the dictionary,
and if it is not - this is a bug.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>