Tpmi: do not use 'ErrorNotFound'
Do not use 'ErrorNotFound' when a bit-field is not found, because it is
inconsistent with 'Error' being used for the case when the register was not
found. Just use 'Error'. If a distinction is needed at some point, it can be
done separately and consistently for all "something requested not found" cases.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>