Tpmi: rename bitname to bfname
I would like to have the "name" part, so that it is clear this is just a name,
not a tuple with bit range, like in MSR.py. But 'bitname' is confusing too.
Rename it to 'bfname' (bit field name). Although this is not ideal name for the
variable either, but it is less confusing than 'bitname'.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>