Tpmi: fix instances range
Use the actual instance numbers in the error message, do not assume they are
sequential and start with 0, there is no reason or need to assume that. Just
use 'mdmap' keys for that - they are the actually found instance numbers.

Improve the error message a bit as well. Plus a random change: available ->
packages variable rename, just for consistency, since all the other occurrences
of this variable were removed too.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>