_PepcTpmi: get rid of '_parse_tpmi_args()'
First of all, it does not seem very readable because of too many return
arguments. Second, does not seem that optimal because it assigned "all" to some
arguments, then the values are changed very soon after it returns.

Also, other sub-command do not use a method like this, just open-code
everything in 'tpmi_read_command()' instead for consistency.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>